キャスト関連の警告に対応

+++ dlginvestigate.cpp
@@ -13,6 +13,7 @@
 #include "vkeytable.h"
 #include "dlginvestigate.h"
 #include <iomanip>
+#include <cstdint>
 
 
 ///
@@ -84,7 +85,7 @@ public:
 				{
 					Acquire a(&m_data.m_engine->m_log, 1);
 					m_data.m_engine->m_log << _T("HWND:\t") << std::hex
-					<< reinterpret_cast<int>(i_hwndTarget)
+					<< static_cast<DWORD>(reinterpret_cast<uintptr_t>(i_hwndTarget))
 					<< std::dec << std::endl;
 				}
 				Acquire a(&m_data.m_engine->m_log, 0);
